Added support for mapping subclasses
The annotation processor now finds the subclasses of all @Mappable classes and adds their fields to the subclasses mappers.
There is a strange behaviour in "processParseAnnotationElements()": The parsers of the subclasses mappers are generated in their place event I comment the new method in charge of doing it (seems like in the following steps they are being added somehow...).